+++ lib/legion/extensions/microsoft_teams/cli/auth.rb
@@ -8,8 +8,8 @@ module Legion
     module MicrosoftTeams
       module CLI
         class Auth
-          include Legion::Extensions::Helpers::Lex if Legion::Extensions.const_defined?(:Helpers) &&
-                                                      Legion::Extensions::Helpers.const_defined?(:Lex)
+          include Legion::Extensions::Helpers::Lex if Legion::Extensions.const_defined?(:Helpers, false) &&
+                                                      Legion::Extensions::Helpers.const_defined?(:Lex, false)
 
           def self.cli_alias
             'teams'