+++ lib/legion/extensions/microsoft_teams/runners/presence.rb
@@ -22,8 +22,8 @@ module Legion
             { availability: 'Offline', activity: 'OffWork', error: e.message, fetched_at: Time.now.utc }
           end
 
-          include Legion::Extensions::Helpers::Lex if Legion::Extensions.const_defined?(:Helpers) &&
-                                                      Legion::Extensions::Helpers.const_defined?(:Lex)
+          include Legion::Extensions::Helpers::Lex if Legion::Extensions.const_defined?(:Helpers, false) &&
+                                                      Legion::Extensions::Helpers.const_defined?(:Lex, false)
         end
       end
     end